Create Identify local accounts
This query looks for specific Security Event IDs (4720, 4722, 4723, 4724, 4725, 4726, 4738, 4739, 4740, and 4741) that are related to local account management. It then filters the results to only show user accounts with an AccountDomain of '$' (indicating a local account) and excludes the 'ANONYMOUS LOGON' account. The query also includes additional fields such as the time the event was generated, the account name, event ID, account type, the computer name, and the target user name.