Create Search for a specific CVE vulnerability
This query uses the SecurityEvent table to search for events related to modifications of registry values that contain a specific CVE identifier. In this example, the query is searching for CVE-2021-1234, but you can replace this value with the CVE identifier you want to search for. The query displays the time of the event, the name of the computer where the event occurred, and the registry value that contains the CVE identifier.

Note that this is just an example query and may need to be modified based on your specific environment and requirements. Additionally, this query assumes that the CVE identifier is included in the registry value and follows the format "CVE-YYYY-NNNN", where YYYY is the year and NNNN is the vulnerability identifier. If the CVE identifier is stored in a different format or location, the query may need to be modified accordingly.