Create Detect Exchange Zero Day vulnerabilities
This script will search the Exchange audit logs for the past 7 days for successful commands that modify mailbox permissions using the Add-MailboxPermission and Remove-MailboxPermission cmdlets. It will then group the results by cmdlet, mailbox or user, and access rights, and count the number of times each combination appears. The script can be customized to search for other types of Exchange Zero Day vulnerabilities by modifying the cmdlets and parameters used in the Search-ExchangeAuditLog cmdlet.