Create Anomalous amount of SMB sessions created
This script searches the SecurityEvent logs for Event ID 5140, which indicates that an SMB session was created. It then filters the results to only include logon type 3 (network) and a specified time range. The script groups the results by source IP address, destination IP address, computer, and time, and counts the number of SMB sessions created in each group. It then calculates the average number of SMB sessions created per group and uses this value to determine the threshold for what is considered anomalous (in this example, any count greater than 100% of the average is considered anomalous). Finally, the script returns all SMB sessions that exceed the threshold.

This script can be customized by adjusting the time range and threshold values to suit the organization's needs. Additionally, other filters and conditions can be added to further refine the results.