Create Windows network sniffing
This script searches the NetworkCommunicationEvents logs for TCP traffic and groups the results by local IP address, remote IP address, local port, remote port, and time. It then counts the number of TCP connections in each group and calculates the average number of TCP connections per group. The script uses this value to determine the threshold for what is considered potentially sniffing (in this example, any count greater than 50% of the average is considered potentially sniffing). Finally, the script joins the results with the DeviceInventory table to filter out non-Windows devices and returns the TCP connections that exceed the threshold, along with their local and remote IP addresses, local and remote ports, and computer name.

This script can be customized by adjusting the time range and threshold values to suit the organization's needs. Additionally, other filters and conditions can be added to further refine the results.