Create Detect Container lateral movement
This script uses the SecurityAlert table to look for alerts generated in the past timeframe (in this case, one day) that match the criteria for suspicious container activity. It filters the results to only show alerts related to the container with the name containerName.

The script then uses the parse_json function to extract information from the TargetAccounts field, which contains details about the accounts used in the suspicious activity. It filters the results to only show alerts where the first account in the TargetAccounts array performed a network connect operation.

Finally, the script projects the relevant information, including the time the alert was generated, the name of the alert, the name of the account used in the suspicious activity, the resource the account accessed, the IP address the account connected to, and the port and protocol used in the network connection.